Flipped the timestep scaling
Before: TicksPerHour was used to subdivide the hourly timesteps into
shorter intervals.
After: HoursPerTick accumulates several hours to a snigle timestep

+++ Agro/Program.cs
@@ -35,7 +35,7 @@ internal class Program
         else
             world = Initialize.World();
         var start = DateTime.UtcNow.Ticks;
-        world.Run(AgroWorld.TimestepsTotal);
+        world.Run((uint)AgroWorld.TimestepsTotal());
         var stop = DateTime.UtcNow.Ticks;
         Console.WriteLine($"Simulation time: {(stop - start) / TimeSpan.TicksPerMillisecond} ms");
 #if HISTORY_LOG || HISTORY_TICK
